+++ ios/Flutter/flutter_export_environment.sh
@@ -5,6 +5,7 @@ export "FLUTTER_APPLICATION_PATH=/Users/billbunkum/Development/githubs/oneChoice
 export "FLUTTER_TARGET=/Users/billbunkum/Development/githubs/oneChoice/lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
+export "OTHER_LDFLAGS=$(inherited) -framework Flutter"
 export "FLUTTER_FRAMEWORK_DIR=/Users/billbunkum/Development/flutter/bin/cache/artifacts/engine/ios"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"
