+++ lib/widgets/somethingGoodForm.dart
@@ -9,88 +9,15 @@ class SomethingGoodForm extends StatefulWidget {
 class _SomethingGoodFormState extends State<SomethingGoodForm> {
   final _formKey = GlobalKey<FormState>();
 
-  List<String> _onegoodthing = [''];
-
-  getFormWidget() {
-    TextFormField(
-      initialValue: null,
-      decoration: const InputDecoration(
-        border: OutlineInputBorder(),
-        hintText: 'The coffee is great at...',
-        labelText: 'One good thing...',
-      ),
-      onSaved: (String value) {
-        setState(() {
-          _onegoodthing.add(value);
-        });
-      },
-      validator: (String value) {
-        return value.contains('@') ? 'The @ char is not allowed.' : null;
-      },
-    );
-  }
+  List<String> _oneGoodThing = [
+    'yessir', 'here\'s some stuff', 'good things here, sukka',
+  ];
+  // List<Map<String, String>> _oneGoodThing;
   
   @override
   Widget build(BuildContext context) {
-    return Form(
-      key: _formKey,
-      // child: getFormWidget(),
-      child: Column(
-        children: <Widget>[
-          getFormWidget(),
-          Text('hello'),
-        ],
-      )
+    return Container(
+      child: Text('\"text field\" part of MyHomePage Body goes here'),
     );
   }
 }
-
-  // @override
-  // Widget build(BuildContext context) {
-  //   return Form(
-  //     key: _formKey,
-  //     child: getFormWidget(), 
-  //   );
-  // }
-
-  // List<Widget> getFormWidget() {
-  //   List<Widget> formWidget = new List();
-    
-  //   formWidget.add(new TextFormField(
-  //     decoration: InputDecoration(labelText: 'What\'s good?', 
-  //       hintText: 'Coffee is good'),
-  //     onSaved: (value) {
-  //       setState(() {
-  //         _name = value;
-  //       });
-  //     },
-  //   ));
-
-  // }
-// }
-
-
-// class SomethingGoodInput2 extends StatelessWidget {
-//   @override
-//   Widget build(BuildContext context) {
-//     return Column(
-//       children: <Widget>[
-//         TextFormField(
-//           initialValue: null,
-//           decoration: const InputDecoration(
-//             border: OutlineInputBorder(),
-//             hintText: 'The coffee is great at...',
-//             labelText: 'One good thing...',
-//           ),
-//           onSaved: (String value) {
-//             // code runs when User saves form
-            
-//           },
-//           validator: (String value) {
-//             return value.contains('@') ? 'The @ char is not allowed.' : null;
-//           },
-//         ),
-//       ],
-//     );
-//   }
-// }
